create custom bulk action class
